+++ lib/src/html/templates.dart
@@ -13,6 +13,7 @@ import 'package:dartdoc/src/third_party/pkg/mustache4dart/lib/mustache4dart.dart
 const _partials = const <String>[
   'callable',
   'callable_multiline',
+  'categorization',
   'class',
   'constant',
   'footer',
@@ -24,6 +25,7 @@ const _partials = const <String>[
   'documentation',
   'name_summary',
   'sidebar_for_class',
+  'sidebar_for_category',
   'sidebar_for_enum',
   'source_code',
   'sidebar_for_library',
@@ -79,6 +81,7 @@ Future<String> _getTemplateFile(String templateFileName) =>
     loader.loadAsString('package:dartdoc/templates/$templateFileName');
 
 class Templates {
+  final TemplateRenderer categoryTemplate;
   final TemplateRenderer classTemplate;
   final TemplateRenderer enumTemplate;
   final TemplateRenderer constantTemplate;
@@ -115,6 +118,7 @@ class Templates {
 
     var indexTemplate = await _loadTemplate('index.html');
     var libraryTemplate = await _loadTemplate('library.html');
+    var categoryTemplate = await _loadTemplate('category.html');
     var classTemplate = await _loadTemplate('class.html');
     var enumTemplate = await _loadTemplate('enum.html');
     var functionTemplate = await _loadTemplate('function.html');
@@ -131,6 +135,7 @@ class Templates {
 
     return new Templates._(
         indexTemplate,
+        categoryTemplate,
         libraryTemplate,
         classTemplate,
         enumTemplate,
@@ -147,6 +152,7 @@ class Templates {
 
   Templates._(
       this.indexTemplate,
+      this.categoryTemplate,
       this.libraryTemplate,
       this.classTemplate,
       this.enumTemplate,